// projects/isy-angular-widgets-demo/src/app/shared/model/forms.ts
export interface PersonalInformation {
  lastName: string;
  birthName: string;
  birthplace: string;
  firstName: string;
  gender: string;
  birthDate: string;
  nationality: string;
  phoneNumber: string;
  dateOfEntry: string;
  idRequired: string;
  securityLevel: string;
  intelligenceNotes: string;
}